ContentTypeAttrs to its own file
